# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MuscleVolumeOut": {
        "properties": {
          "name_tr": {
            "title": "Name Tr",
            "type": "string"
          },
          "region": {
            "title": "Region",
            "type": "string"
          },
          "sets": {
            "title": "Sets",
            "type": "number"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "svg_id": {
            "title": "Svg Id",
            "type": "string"
          },
          "target": {
            "title": "Target",
            "type": "integer"
          }
        },
        "required": [
          "slug",
          "name_tr",
          "svg_id",
          "region",
          "sets",
          "target"
        ],
        "title": "MuscleVolumeOut",
        "type": "object"
      },
...
      "ProgressionOut": {
        "properties": {
          "alternative_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alternative Label"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "plateau_sessions": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plateau Sessions"
          },
          "reps": {
            "title": "Reps",
            "type": "integer"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          },
          "weight_kg": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Weight Kg",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "weight_kg",
          "reps",
          "label",
          "message"
        ],
        "title": "ProgressionOut",
        "type": "object"
      },
      "RecordOut": {
        "properties": {
          "exercise_id": {
            "format": "uuid",
            "title": "Exercise Id",
            "type": "string"
          },
          "reps": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reps"
          },
          "type": {
            "$ref": "#/components/schemas/PRType"
          },
          "value": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "exercise_id",
          "type",
          "value",
          "reps"
        ],
        "title": "RecordOut",
        "type": "object"
      },
      "SessionOut": {
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_deload": {
            "title": "Is Deload",
            "type": "boolean"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "program_day_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Program Day Id"
          },
          "sets": {
            "items": {
              "$ref": "#/components/schemas/SetOut"
            },
            "title": "Sets",
            "type": "array"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "program_day_id",
          "started_at",
          "completed_at",
          "notes",
          "is_deload"
        ],
        "title": "SessionOut",
        "type": "object"
      },
      "SessionStart": {
        "properties": {
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "program_day_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Program Day Id"
          }
        },
        "title": "SessionStart",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/jwt/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "Progresif overload merkezli antrenman, beslenme ve sağlık takip sistemi.",
    "title": "overload API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/workouts/muscle-volume": {
      "get": {
        "operationId": "muscle_volume_workouts_muscle_volume_get",
        "parameters": [
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 7,
              "maximum": 90,
              "minimum": 1,
              "title": "Days",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MuscleVolumeOut"
                  },
                  "title": "Response Muscle Volume Workouts Muscle Volume Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Muscle Volume",
        "tags": [
          "workouts"
        ]
      }
    },
    "/workouts/progression/{exercise_id}": {
      "get": {
        "operationId": "exercise_progression_workouts_progression__exercise_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "exercise_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Exercise Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Exercise Progression",
        "tags": [
          "workouts"
        ]
      }
    },
    "/workouts/records": {
      "get": {
        "operationId": "list_records_workouts_records_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RecordOut"
                  },
                  "title": "Response List Records Workouts Records Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Records",
        "tags": [
          "workouts"
        ]
      }
    },
    "/workouts/sessions": {
      "get": {
        "operationId": "list_sessions_workouts_sessions_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SessionOut"
                  },
                  "title": "Response List Sessions Workouts Sessions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Sessions",
        "tags": [
          "workouts"
        ]
      },
      "post": {
        "description": "Yeni seans başlatır.\n\nYarım kalmış bir seans varsa yenisi açılmaz — kullanıcı ya onu bitirmeli ya\nda silmeli. İki açık seans, set kayıtlarının hangisine gideceğini belirsiz\nyapar ve ilerleme geçmişini bozar.",
        "operationId": "start_session_workouts_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionStart"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Start Session",
        "tags": [
          "workouts"
        ]
      }
    },
    "/workouts/sessions/{session_id}": {
      "delete": {
        "operationId": "delete_session_workouts_sessions__session_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Delete Session",
        "tags": [
          "workouts"
        ]
      },
      "get": {
        "operationId": "get_session_workouts_sessions__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Session",
        "tags": [
          "workouts"
        ]
      }
    },
    "/workouts/sessions/{session_id}/complete": {
      "post": {
        "description": "Seansı kapatır ve kırılan rekorları tespit eder.\n\nRekor tespiti kapanışta yapılıyor, her set kaydında değil: seans hacmi\nrekoru ancak tüm setler girildikten sonra hesaplanabilir.",
        "operationId": "complete_session_workouts_sessions__session_id__complete_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompleteOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Complete Session",
        "tags": [
          "workouts"
        ]
      }
    },
    "/workouts/sessions/{session_id}/sets": {
      "post": {
        "description": "Tek set kaydeder. Aynı slot tekrar gönderilirse üzerine yazılır —\nkullanıcı yanlış girip düzeltmek isteyebilir.",
        "operationId": "log_set_workouts_sessions__session_id__sets_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Log Set",
        "tags": [
          "workouts"
        ]
      }
    },
    "/workouts/sets/{set_id}": {
      "delete": {
        "operationId": "delete_set_workouts_sets__set_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "set_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Set Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Delete Set",
        "tags": [
          "workouts"
        ]
      }
    },
    "/workouts/streak": {
      "get": {
        "operationId": "streak_workouts_streak_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StreakOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Streak",
        "tags": [
          "workouts"
        ]
      }
    },
    "/workouts/today": {
      "get": {
        "description": "Bugünkü planlanan antrenman, her hareket için ilerleme önerisiyle.\n\nHangi günün sırada olduğu takvim gününe göre DEĞİL, tamamlanan seans\nsayısına göre belirlenir. Sebep: kullanıcı bir günü kaçırırsa programın\ntamamı kaymamalı, kaldığı yerden devam etmeli.",
        "operationId": "todays_workout_workouts_today_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodayOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Todays Workout",
        "tags": [
          "workouts"
        ]
      }
    }
  }
}